AI资讯/ai-news-aggregator/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
feedparser>=6.0.0
python-dateutil>=2.8.0
playwright>=1.58.0
cloudscraper>=1.2.71
